+++ src/app/mypage/page.tsx
@@ -10,49 +10,24 @@ import ModalView from '@/components/organisms/modal/modalView';
 import Login from '@/components/organisms/modal/Login';
 import Script from 'next/script';
 import useUser from '@/hooks/useUser';
+import OnlyUser from '@/components/templates/login/OnlyUser';
 
 const Page = () => {
   const [toggle, setToggle] = useState(0);
-
-  const { user, isLoading: isUserLoading, isError: isUserError } = useUser();
-
-  const loginFn = () => {
-    window.Kakao.Auth.authorize({
-      redirectUri: process.env.NODE_ENV === 'development' ? KAKAO_REDIRECT_URI_DEVELOPMENT : KAKAO_REDIRECT_URI_DEPLOY,
-    });
-  };
 
   const toggleFn = (toggleId: number) => {
     setToggle(toggleId);
   };
 
-  console.log(user);
   return (
-    <>
-      {user && (
-        <div className='flex flex-col items-center justify-start min-h-screen min-w-full'>
-          <div className=''>
-            <StudyToggle activeToggle={toggle} toggleFn={toggleFn} />
-          </div>
-          <div className='mt-63'>{toggle === 0 ? <Favorites /> : <Setting />}</div>
+    <OnlyUser>
+      <div className='flex flex-col items-center justify-start min-h-screen min-w-full'>
+        <div className=''>
+          <StudyToggle activeToggle={toggle} toggleFn={toggleFn} />
         </div>
-      )}
-      {isUserError && (
-        <>
-          <BackDrop>
-            <ModalView>
-              <Login loginFn={loginFn} closeFn={() => {}} />
-            </ModalView>
-          </BackDrop>
-          <Script
-            src='https://developers.kakao.com/sdk/js/kakao.js'
-            onLoad={() => {
-              window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_JS_KEY);
-            }}
-          />
-        </>
-      )}
-    </>
+        <div className='mt-63'>{toggle === 0 ? <Favorites /> : <Setting />}</div>
+      </div>
+    </OnlyUser>
   );
 };
