+++ src/constants/localhostKeys.ts
@@ -1,0 +1,5 @@
+const LOCALHOST_KEYS = {
+  redirectUrl: 'redirectUrl',
+};
+
+export default LOCALHOST_KEYS;
